Minor refactoring, moving things around
