+++ src/mod.rs
@@ -2,8 +2,10 @@ mod piece_table;
 mod workspace;
 mod cursor;
 mod editor;
+mod editor_element;
 
 pub mod workspace;
 pub mod piece_table;
 pub mod cursor;
 pub mod editor;
+pub mod editor_element;
